# openapi.yaml
...
openapi: 3.0.0
info:
  x-skaffolder-id-project: 5dc0d95e9ef7a1181f8e1322
  title: khachsan
  description: You can obtain JWT token from login API
  version: 0.1
servers:
  - url: http://localhost:3000/api
    description: Local server for testing
paths:
  /login:
    post:
      x-skaffolder-ignore: true
      summary: Login API
      security: []   # No security
      tags:
        - Security
      requestBody:
        content:
          application/json:
            schema:      # Request body contents
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
              example:   # Sample object
                username: admin
                password: 62f264d7ad826f02a8af714c0a54b197935b717656b80461686d450f7b3abde4c553541515de2052b9af70f710f0cd8a1a2d3f4d60aa72608d71a63a9a93c0f5
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /verifyToken:
    post:
      x-skaffolder-ignore: true
      summary: Verify token API
      security: [] #No security
      tags:
        - Security
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                token:
                  type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /changePassword:
    post:
      x-skaffolder-ignore: true
      summary: Change password API
      tags:
        - Security
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                passwordNew:
                  type: string
                passwordOld:
                  type: string
      responses:
        '200':
          description: OK
                

  #-----  User  -----
...
  /user/:
    post:
      x-skaffolder-id: 5dc0d95e9ef7a1181f8e1334
      x-skaffolder-name: create
      x-skaffolder-id-resource: 5dc0d95e9ef7a1181f8e132c
      x-skaffolder-resource: User
      x-skaffolder-crudAction: create
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION create
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION create
      tags:
        - User
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                employess:
                  type: integer
                mail:
                  type: string
                name:
                  type: string
                password:
                  type: string
                roles:
                  type: string
                surname:
                  type: string
                username:
                  type: string
    get:
      x-skaffolder-id: 5dc0d95e9ef7a1181f8e1330
      x-skaffolder-name: list
      x-skaffolder-id-resource: 5dc0d95e9ef7a1181f8e132c
      x-skaffolder-resource: User
      x-skaffolder-crudAction: list
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION list
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION list
      tags:
        - User
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
  /user/{id}:
    delete:
      x-skaffolder-id: 5dc0d95e9ef7a1181f8e1336
      x-skaffolder-name: delete
      x-skaffolder-id-resource: 5dc0d95e9ef7a1181f8e132c
      x-skaffolder-resource: User
      x-skaffolder-crudAction: delete
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION delete
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION delete
      tags:
        - User
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      x-skaffolder-id: 5dc0d95e9ef7a1181f8e1332
      x-skaffolder-name: get
      x-skaffolder-id-resource: 5dc0d95e9ef7a1181f8e132c
      x-skaffolder-resource: User
      x-skaffolder-crudAction: get
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION get
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION get
      tags:
        - User
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    post:
      x-skaffolder-id: 5dc0d95e9ef7a1181f8e132f
      x-skaffolder-name: update
      x-skaffolder-id-resource: 5dc0d95e9ef7a1181f8e132c
      x-skaffolder-resource: User
      x-skaffolder-crudAction: update
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION update
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION update
      tags:
        - User
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
    
  #-----  bill  -----
  /bill/:
    post:
      x-skaffolder-id: 5dc0ef119ef7a1181f8e1470
      x-skaffolder-name: create
      x-skaffolder-id-resource: 5dc0e8d59ef7a1181f8e13fb
      x-skaffolder-resource: bill
      x-skaffolder-crudAction: create
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION create
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION create
      tags:
        - bill
      parameters:
        - name: obj
          x-skaffolder-type: bill
          in: path
          description: Object to insert
          required: true
          schema:
            type: bill
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/bill"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                date:
                  type: integer
                employessid:
                  type: integer
                roomid:
                  type: integer
                salarybill:
                  type: number
    get:
      x-skaffolder-id: 5dc0eeba9ef7a1181f8e1469
      x-skaffolder-name: list
      x-skaffolder-id-resource: 5dc0e8d59ef7a1181f8e13fb
      x-skaffolder-resource: bill
      x-skaffolder-crudAction: list
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION list
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: ARRAY OF bill
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION list
      tags:
        - bill
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/bill"
  /bill/{id}:
    delete:
      x-skaffolder-id: 5dc0ef0b9ef7a1181f8e146e
      x-skaffolder-name: delete
      x-skaffolder-id-resource: 5dc0e8d59ef7a1181f8e13fb
      x-skaffolder-resource: bill
      x-skaffolder-crudAction: delete
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION delete
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION delete
      tags:
        - bill
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id bill
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      x-skaffolder-id: 5dc0ef199ef7a1181f8e1472
      x-skaffolder-name: get
      x-skaffolder-id-resource: 5dc0e8d59ef7a1181f8e13fb
      x-skaffolder-resource: bill
      x-skaffolder-crudAction: get
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION get
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: bill
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION get
      tags:
        - bill
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id bill
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/bill"
    post:
      x-skaffolder-id: 5dc0ef049ef7a1181f8e146c
      x-skaffolder-name: update
      x-skaffolder-id-resource: 5dc0e8d59ef7a1181f8e13fb
      x-skaffolder-resource: bill
      x-skaffolder-crudAction: update
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION update
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: bill
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION update
      tags:
        - bill
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id bill
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/bill"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/bill"
  /bill/findBydate/{key}:
    get:
      x-skaffolder-id: 5dc0eea99ef7a1181f8e1467
      x-skaffolder-name: findBydate
      x-skaffolder-id-resource: 5dc0e8d59ef7a1181f8e13fb
      x-skaffolder-resource: bill
      x-skaffolder-crudAction: findBydate
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION findBydate
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /findBydate/{key}
      x-skaffolder-roles: []
      summary: CRUD ACTION findBydate
      tags:
        - bill
      parameters:
        - name: key
          x-skaffolder-type: Objectid
          in: path
          description: Id of the resource date to search
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/bill"
    
  #-----  costomer  -----
...
security:
  - bearerAuth: [] 
  
  
# Components
components:
...
  securitySchemes:
    bearerAuth:            # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT    # optional, arbitrary value for documentation purposes

  # Data Model
  schemas:
    User:
      x-skaffolder-id: 5dc0d95e9ef7a1181f8e132c
      x-skaffolder-id-db: 5dc0d95e9ef7a1181f8e1323
      x-skaffolder-id-entity: 5dc0d95e9ef7a1181f8e1324
      x-skaffolder-url: /user
      properties:
        _id:
          type: string
          x-skaffolder-required: true
        employess:
          type: integer
          x-skaffolder-id-attr: 5dc0eb1d9ef7a1181f8e1413
          x-skaffolder-type: Integer
        mail:
          type: string
          x-skaffolder-id-attr: 5dc0d95e9ef7a1181f8e1327
          x-skaffolder-type: String
        name:
          type: string
          x-skaffolder-id-attr: 5dc0d95e9ef7a1181f8e1328
          x-skaffolder-type: String
        password:
          type: string
          x-skaffolder-id-attr: 5dc0d95e9ef7a1181f8e1326
          x-skaffolder-type: String
          x-skaffolder-required: true
        roles:
          type: string
          x-skaffolder-id-attr: 5dc0d95e9ef7a1181f8e132a
          x-skaffolder-type: String
        surname:
          type: string
          x-skaffolder-id-attr: 5dc0d95e9ef7a1181f8e1329
          x-skaffolder-type: String
        username:
          type: string
          x-skaffolder-id-attr: 5dc0d95e9ef7a1181f8e1325
          x-skaffolder-type: String
          x-skaffolder-required: true
      x-skaffolder-relations: 
      required:
        - _id
        - password
        - username
    bill:
      x-skaffolder-id: 5dc0e8d59ef7a1181f8e13fb
      x-skaffolder-id-db: 5dc0d95e9ef7a1181f8e1323
      x-skaffolder-id-entity: 5dc0e8d59ef7a1181f8e13f6
      x-skaffolder-url: /bill
      properties:
        _id:
          type: string
          x-skaffolder-required: true
        date:
          type: integer
          x-skaffolder-id-attr: 5dc0e8d59ef7a1181f8e13f9
          x-skaffolder-type: Date
        employessid:
          type: integer
          x-skaffolder-id-attr: 5dc0e8d59ef7a1181f8e13f7
          x-skaffolder-type: Integer
        roomid:
          type: integer
          x-skaffolder-id-attr: 5dc0e8d59ef7a1181f8e13f8
          x-skaffolder-type: Integer
        salarybill:
          type: number
          x-skaffolder-id-attr: 5dc0e8d59ef7a1181f8e13fa
          x-skaffolder-type: Decimal
      x-skaffolder-relations: 
      required:
        - _id
...